Refactor into class for usage in pyopenms
