+++ src/openms/configh.cmake
@@ -211,6 +211,6 @@ add_custom_command (
 )
 
 #------------------------------------------------------------------------------
-# export a list of all configured heders
+# export a list of all configured headers
 set(OpenMS_configured_headers "${CONFIGURED_BUILD_CONFIG_CURRENT_H};${CONFIGURED_CONFIG_H};${CONFIGURED_OPENMS_PACKAGE_VERSION_H};${CONFIGURED_OPENMS_DATA_PATH_H}")
 set_property(SOURCE ${CONFIGURED_BUILD_CONFIG_CURRENT_H} PROPERTY SKIP_AUTOMOC ON)
