+++ src/openms/include/OpenMS/SYSTEM/sources.cmake
@@ -19,6 +19,7 @@ source_group("Source Files\\OpenMS\\SYSTEM" FILES ${sources})
 
 ### list all header files of the directory here
 set(sources_list_h
+BuildInfo.h
 ExternalProcess.h
 File.h
 FileWatcher.h
