+++ src/utils/OpenMSInfo.cpp
@@ -33,172 +33,20 @@
 // --------------------------------------------------------------------------
 
 #include <OpenMS/config.h>
-#include <OpenMS/build_config.h>
 #include <OpenMS/CONCEPT/VersionInfo.h>
 #include <OpenMS/DATASTRUCTURES/String.h>
+#include <OpenMS/SYSTEM/BuildInfo.h>
 #include <OpenMS/SYSTEM/File.h>
 #include <OpenMS/openms_data_path.h>
 
-#include <QSysInfo>
-#include <QDir>
-
 #ifdef _OPENMP
   #include "omp.h"
-#endif
-
-#if defined(Q_WS_MAC)
-  #include <CoreServices/CoreServices.h>
 #endif
 
 #include <iostream>
 
 using namespace OpenMS;
 using namespace std;
-
-namespace OpenMS
-{
-  namespace Internal
-  {
-
-    enum OpenMS_OS {OS_UNKNOWN, OS_MACOS, OS_WINDOWS, OS_LINUX};
-    std::string OpenMS_OSNames[] = {"unknown", "MacOS", "Windows", "Linux"};
-    enum OpenMS_Architecture {ARCH_UNKNOWN, ARCH_32BIT, ARCH_64BIT};
-    std::string OpenMS_ArchNames[] = {"unknown", "32 bit", "64 bit"};
-
-#if WIN32
-    OpenMS_Architecture getArchOnWin();
-    String getWinOSVersion();
-#endif
-
-    struct OpenMSOSInfo
-    {
-      OpenMSOSInfo() :
-        os(OS_UNKNOWN),
-        os_version("unknown"),
-        arch(ARCH_UNKNOWN)
-      {}
-
-      OpenMS_OS os;
-      String os_version;
-      OpenMS_Architecture arch;
-
-      String getOSAsString() const
-      {
-        return OpenMS_OSNames[os];
-      }
-
-      String getArchAsString() const
-      {
-        return OpenMS_ArchNames[arch];
-      }
-
-    };
-
-    static OpenMSOSInfo getOSInfo()
-    {
-      OpenMSOSInfo info;
-#if defined(WIN32)  // Windows
-      info.os = OS_WINDOWS;
-      info.arch = getArchOnWin();
-      info.os_version = getWinOSVersion();
-#elif (defined(__MACH__) && defined(__APPLE__)) // MacOS
-      info.os = OS_MACOS;
-#else //Linux
-      info.os = OS_LINUX;
-#endif
-       // check if we can use QSysInfo
-#if (defined(Q_OS_MACOS) || defined(Q_OS_UNIX))
-
-      info.os_version = QSysInfo::productVersion();
-
-      // identify architecture
-      if (QSysInfo::WordSize == 32)
-      {
-        info.arch = ARCH_32BIT;
-      }
-      else
-      {
-        info.arch = ARCH_64BIT;
-      }
-#endif
-
-      return info;
-    }
-
-//********************
-//  Windows specific API calls
-//********************
-#ifdef WIN32
-#include <windows.h>
-#include <stdio.h>
-
-    typedef BOOL (WINAPI * LPFN_ISWOW64PROCESS)(HANDLE, PBOOL);
-
-    LPFN_ISWOW64PROCESS fnIsWow64Process;
-
-    OpenMS_Architecture getArchOnWin()
-    {
-#ifdef OPENMS_64BIT_ARCHITECTURE
-      return ARCH_64BIT;
-
-#else
-      BOOL bIsWow64 = FALSE;
-
-      //IsWow64Process is not available on all supported versions of Windows.
-      //Use GetModuleHandle to get a handle to the DLL that contains the function
-      //and GetProcAddress to get a pointer to the function if available.
-
-      fnIsWow64Process = (LPFN_ISWOW64PROCESS) GetProcAddress(
-        GetModuleHandle(TEXT("kernel32")), "IsWow64Process");
-
-      if (NULL != fnIsWow64Process)
-      {
-        if (!fnIsWow64Process(GetCurrentProcess(), &bIsWow64))
-        {
-          return ARCH_UNKNOWN;
-        }
-      }
-      if (bIsWow64)
-      {
-        return ARCH_64BIT;
-      }
-      else
-      {
-        return ARCH_32BIT;
-      }
-#endif
-    }
-
-    String getWinOSVersion()
-    {
-      OSVERSIONINFO osvi;
-      ZeroMemory(&osvi, sizeof(OSVERSIONINFO));
-      osvi.dwOSVersionInfoSize = sizeof(OSVERSIONINFO);
-      GetVersionEx(&osvi);
-      return String(osvi.dwMajorVersion) + "." + String(osvi.dwMinorVersion);
-    }
-
-#endif // WIN32 API functions
-
-    /*
-      @brief Get Architecture of this binary (simply by looking at size of a pointer, i.e. size_t).
-    */
-    OpenMS_Architecture getBinaryArchitecture()
-    {
-      size_t bytes = sizeof(size_t);
-      switch (bytes)
-      {
-        case 4:
-          return ARCH_32BIT;
-        case 8:
-          return ARCH_64BIT;
-        default:
-          return ARCH_UNKNOWN;
-      }
-    }
-
-  } // NS Internal
-} // NS OpenMS
 
 // We do not want this class to show up in the docu:
 /// @cond TOPPCLASSES
@@ -223,21 +71,21 @@ int main(int /*argc*/, const char ** /*argv*/)
   cout << "==================" << "\n";
   cout << "Source path  : " << OPENMS_SOURCE_PATH << "\n";
   cout << "Binary path  : " << OPENMS_BINARY_PATH << "\n";
-  cout << "Binary arch  : " << Internal::OpenMS_ArchNames[Internal::getBinaryArchitecture()] << "\n";
-  cout << "Build type   : " << OPENMS_BUILD_TYPE << "\n";
+  cout << "Binary arch  : " << Internal::getBinaryArchitecture() << "\n";
+  cout << "Build type   : " << Internal::getBuildType() << "\n";
   #ifdef _OPENMP
-  cout << "OpenMP       : " << "enabled (maxThreads = " << omp_get_max_threads() << ")" << "\n";
+  cout << "OpenMP       : " << "enabled (maxThreads = " << Internal::getOpenMPMaxNumThreads() << ")" << "\n";
   #else
   cout << "OpenMP       : " << "disabled" << "\n";
   #endif
   cout << "\n";
 
-  OpenMS::Internal::OpenMSOSInfo info = OpenMS::Internal::getOSInfo();
-  // experimental: OS information
+  Internal::OpenMSOSInfo info = Internal::OpenMSOSInfo::getOSInfo();
+
   cout << "OS Information:" << "\n";
   cout << "==================" << "\n";
   cout << "Name: " << info.getOSAsString() << "\n";
-  cout << "Version: " << info.os_version << "\n";
+  cout << "Version: " << info.getOSVersionAsString() << "\n";
   cout << "Architecture: " << info.getArchAsString() << "\n";
   cout << "\n";
